% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\name{CDEResult}
\alias{CDEResult}
\title{Constructor for CDEResult}
\usage{
CDEResult(deTable, subtypeProportions, subtypeDE, params = list())
}
\arguments{
\item{deTable}{A \code{DataFrame} of per-gene compound DE statistics.}

\item{subtypeProportions}{A \code{matrix} of subtype proportions
(samples x subtypes).}

\item{subtypeDE}{A named \code{list} of per-subtype DE \code{DataFrame}s.}

\item{params}{A \code{list} of analysis parameters.}
}
\value{
A \code{CDEResult} object.
}
\description{
Create a new \code{CDEResult} object.
}
\examples{
library(S4Vectors)
dt <- DataFrame(
    gene      = c("G1", "G2"),
    logFC     = c(1.2, 0.1),
    AveExpr   = c(3.1, 2.0),
    t         = c(4.1, 0.5),
    P.Value   = c(0.001, 0.8),
    adj.P.Val = c(0.01, 0.9),
    B         = c(2.1, -2.0),
    T_score   = c(1.1, 0.05),
    C_score   = c(0.1, 0.09),
    T_score_z = c(1.5, 0.2),
    C_score_z = c(0.2, 0.3),
    TC_ratio  = c(0.88, 0.40),
    source    = c("transcriptional", "mixed")
)
pm <- matrix(c(0.6, 0.4, 0.3, 0.7), nrow = 2,
             dimnames = list(c("D1___ctrl","D1___treat"),
                             c("TypeA","TypeB")))
obj <- CDEResult(deTable = dt, subtypeProportions = pm,
                 subtypeDE = list(), params = list(broad_type = "T_cell"))
obj

}
