% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\name{subtypeDE}
\alias{subtypeDE}
\alias{subtypeDE,CDEResult-method}
\title{Accessor for per-subtype DE results in a CDEResult}
\usage{
subtypeDE(x, ...)

\S4method{subtypeDE}{CDEResult}(x, ...)
}
\arguments{
\item{x}{A \code{CDEResult} object.}

\item{...}{Additional arguments (not used).}
}
\value{
A named \code{list} of \code{DataFrame}s, one per subtype.
}
\description{
Returns the list of per-subtype limma DE \code{DataFrame}s.
}
\examples{
library(S4Vectors)
dt <- DataFrame(gene = "G1", logFC = 1.2, P.Value = 0.001,
                adj.P.Val = 0.01, AveExpr = 3.1, t = 4.1, B = 2.1,
                T_score = 1.1, C_score = 0.1,
                T_score_z = 1.5, C_score_z = 0.2,
                TC_ratio = 0.88, source = "transcriptional")
pm <- matrix(c(0.6, 0.4), nrow = 1,
             dimnames = list("D1___ctrl", c("TypeA", "TypeB")))
obj <- CDEResult(dt, pm, list(), list(broad_type = "T_cell"))
subtypeDE(obj)

}
