% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\name{subtypeProportions}
\alias{subtypeProportions}
\alias{subtypeProportions,CDEResult-method}
\title{Accessor for subtype proportions in a CDEResult}
\usage{
subtypeProportions(x, ...)

\S4method{subtypeProportions}{CDEResult}(x, ...)
}
\arguments{
\item{x}{A \code{CDEResult} object.}

\item{...}{Additional arguments (not used).}
}
\value{
A \code{matrix} with rows = samples and columns = subtypes.
}
\description{
Returns the subtype proportion matrix from a
  \code{CDEResult} object.
}
\examples{
library(S4Vectors)
dt <- DataFrame(gene = "G1", logFC = 1.2, P.Value = 0.001,
                adj.P.Val = 0.01, AveExpr = 3.1, t = 4.1, B = 2.1,
                T_score = 1.1, C_score = 0.1,
                T_score_z = 1.5, C_score_z = 0.2,
                TC_ratio = 0.88, source = "transcriptional")
pm <- matrix(c(0.6, 0.4), nrow = 1,
             dimnames = list("D1___ctrl", c("TypeA", "TypeB")))
obj <- CDEResult(dt, pm, list(), list(broad_type = "T_cell"))
subtypeProportions(obj)

}
